refactor: deduplicate files, merge metrics, remove interfaces
- Delete duplicate cli/inspect/ files (byte-identical copies of
  meta_analysis/ originals), redirect imports.
- Merge metrics_alignment.py into unified_metrics.py.
- Delete plugins/interfaces.py, replace IScanner/IConverter/IReporter
  sentinels with string literals in plugin_manager, execution_engine,
  and cli/report.
- Move test_error_handling.py from production code to tests/.
